requirements.txt
anthropic
streamlit
requests
beautifulsoup4
python-dotenv
pdfplumber
python-docx
lxml
pathlib
